Added two new functions and new conditions
The two functions are in charge of handling the votes submitted by the user. The first function update the choice list related to an active voting question. The second function gets the choices when a voting question is closed. The new conditions are to store the choices of an active voting question in a list once its load and to empty the list when the voting question is closed